Add nav header and set root_path
Runs `rails g refills:import centered_navigation` to pull in some
default styles for the app's header. Changes the links to include the
`books_path` and links for clearance.

The only thing that I find weird is that `HighVoltage` sets the root
route in an initialize, rather than in `routes.rb` which might be
confusing for someone coming onto the project.

https://github.com/thoughtbot/high_voltage#specifying-a-root-path

Still to do: remove the default logo, add some content to
`home.html.erb`.