Add Books#show page
https://trello.com/c/Zqj5vM5w/9-book-page

Completes the first item on the checklist: a show page with the book
title, description and image. I wrote the feature spec to start from the
`root_path` since adding those links is part of the spec for this
project.

Also added a default description to the `books` factory and some basic
styling for `books/show.html.erb`.