refactors for readability and speed
This commit adds small style changes along with creating a readable
transient in our factory for lists. It also adds foreign keys to the
book_lists table and removes the dependent: :destroy from the models.
Our thoughts on removing this were to let Postgres handle referential
integrity in following this guide.
https://robots.thoughtbot.com/referential-integrity-with-foreign-keys

We also return an array in our list model rather than querying the whole
db and then calling a limit to avoid n+1 queries.