Add book review feature
https://trello.com/c/dMKJKgY6/27-book-reviews

Users who are signed in can add a book review if they haven't reviewed
the book yet.
Unauthenticated users cannot review books.

I'm not particularly happy with how the `user.can_review_book?(book)`
method works, it makes an extra db call and I'm not sure if User is the
best place for that logic.