Add basics styles to authors, books, lists, reviews and user forms
Adds a basic container to each page so that the content is somewhat
centered. I opted not to use `rails g clearance:views` and style them,
but all of the signin/signup stuff is wrapped in `#clearance` so we can
style the views without importing them.

Also sets the image size when using `book_cover` and adds an hr in a few
places to make the content a little more readable.