adds sort by average review
https://trello.com/c/XctSVpOz/21-books-sortable-by-date-and-avg-review

This commit adds a sort by average rating to the books index.  We also
added a background job to catch for the case of a large amount of
reviews, and we wouldn't want to delay the render of the page while that
calculates.  This commit also tests the background job and the
controller that calls the background job to see that the method
update_avg_rating is called.